core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Campus Farm EMS real-time control loop"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.0",
    "requests>=2.32",
    "websockets>=16.0",
    "matplotlib>=3.9",
    "numpy>=2.0",
    "pandas>=2.2",
]
